Move `Json` and `Decoder` failures into `json.error` packages
