+++ sbt-github/src/main/scala/com/alejandrohdezma/sbt/github/http/error/URLNotFound.scala
@@ -1,0 +1,5 @@
+package com.alejandrohdezma.sbt.github.http.error
+
+import scala.util.control.NoStackTrace
+
+final case class URLNotFound(url: String) extends Throwable(s"$url was not found") with NoStackTrace
