+++ sbt-github/src/main/scala/com/alejandrohdezma/sbt/github/json/error/NotAJSONObject.scala
@@ -1,0 +1,9 @@
+package com.alejandrohdezma.sbt.github.json.error
+
+import scala.util.control.NoStackTrace
+
+import com.alejandrohdezma.sbt.github.json.Json
+
+final case class NotAJSONObject(value: Json.Value)
+    extends Throwable(s"is not a valid JSON object: $value")
+    with NoStackTrace
