+++ sbt-github/src/main/scala/com/alejandrohdezma/sbt/github/json/error/NotANumber.scala
@@ -1,0 +1,9 @@
+package com.alejandrohdezma.sbt.github.json.error
+
+import scala.util.control.NoStackTrace
+
+import com.alejandrohdezma.sbt.github.json.Json
+
+final case class NotANumber(value: Json.Value)
+    extends Throwable(s"is not a valid JSON number: $value")
+    with NoStackTrace
