+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/http/ClientSpec.scala
@@ -24,7 +24,7 @@ import cats.implicits._
 
 import sbt.util.Logger
 
-import com.alejandrohdezma.sbt.github.http.client.URLNotFound
+import com.alejandrohdezma.sbt.github.http.error.URLNotFound
 import com.alejandrohdezma.sbt.github.json.Decoder
 import com.alejandrohdezma.sbt.github.syntax.json.JsonValueOps
 import com.alejandrohdezma.sbt.github.withServer
