+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/syntax/JsonSyntaxSpec.scala
@@ -17,9 +17,8 @@
 package com.alejandrohdezma.sbt.github.syntax
 
 import com.alejandrohdezma.sbt.github.error._
-import com.alejandrohdezma.sbt.github.json.Decoder.Failure.NotABoolean
 import com.alejandrohdezma.sbt.github.json.Json
-import com.alejandrohdezma.sbt.github.json.Json.Failures.{InvalidPath, NotAJSONObject}
+import com.alejandrohdezma.sbt.github.json.error._
 import com.alejandrohdezma.sbt.github.syntax.json._
 import org.specs2.mutable.Specification
